experimental filtering of delta events
    delta gesture -> filter -> model;
    button push gesture -> filter;

When an Arc's push button is pressed, it's quite easy to mess up with
one's value when pressing/releasing (an rotary movement gets recorded
sometimes)

We attempt to solve this using a filter around the time the button was
pressed. Events produced afterwards are ignored. Events received just
before are rolled-back.